slack-grounding-recommendations/requirements.txt
google-cloud-functions>=1.8.0
google-cloud-secret-manager>=2.16.0
notion-client>=2.0.0
slack-sdk>=3.21.0
anthropic>=0.3.0
python-dotenv>=1.0.0
google-cloud-scheduler>=2.12.0
functions-framework>=3.4.0 